Add files via upload
Before we use Keras deep learning framework, we try to understand how to program neural network by purely using python. This is to get a mathematical understanding of deep learning. We define a network with 1 hidden layer having 2 neurons and 1 output layer. The input to the network is different combinations of 0,1 and the output is logical XOR result of the given combination. Each neuron has different weights and biases, and are updated through different epochs of training. We define custom functions for forward pass, activation, classification, and weight updation.